src/main/webapp/app/entities/log-dang-nhap/delete/log-dang-nhap-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { provideHttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { LogDangNhapService } from '../service/log-dang-nhap.service';

import { LogDangNhapDeleteDialogComponent } from './log-dang-nhap-delete-dialog.component';

describe('LogDangNhap Management Delete Component', () => {
  let comp: LogDangNhapDeleteDialogComponent;
  let fixture: ComponentFixture<LogDangNhapDeleteDialogComponent>;
  let service: LogDangNhapService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [LogDangNhapDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(LogDangNhapDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(LogDangNhapDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(LogDangNhapService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
